Typ der Arraylist auf Integer

+++ SE_Grp73/src/se_ex01/KIRandom.java
@@ -15,7 +15,7 @@ public class KIRandom extends KI {
 	public int chooseNumber() {
 
 		int randomPosition = (int) (Math.random() * (remainingNumbers.size()-1));
-		int chosenNumber = (int) remainingNumbers.get(randomPosition);
+		int chosenNumber = remainingNumbers.get(randomPosition);
 		remainingNumbers.remove(chosenNumber);
 		return chosenNumber;
